Issue #276: extract inline route logic to service methods
Create 7 new service files following the lazy singleton factory pattern:
- service-error.ts: Base ServiceError class with HTTP status code mapping
- message-template-service.ts: listTemplates(), upsertTemplate()
- pr-service.ts: enableAutoMerge(), disableAutoMerge(), updateBranch()
- diagnostics-service.ts: getBlockedTeams(), getHealthSummary(), factoryReset()
- project-service.ts: listProjects(), createProject(), getProjectDetail(),
  deleteProject(), installHooksForProject(), getPrompt(), savePrompt()
- project-group-service.ts: listWithCounts(), getWithProjects()
- team-service.ts: launchTeam(), launchBatch(), getTeamDetail(), getTeamStatus(),
  getTeamTimeline(), exportTeam(), sendMessage(), setPhase(), acknowledgeAlert()

Update all 7 route files to delegate business logic to service methods:
routes become thin HTTP handlers (parse params, call service, send response).
Services own all DB calls, SSE broadcasts, filesystem ops, and CLI calls.

Pure refactoring: no API behavior changes, no new endpoints, no deleted endpoints.
MCP tools can now call the same service methods without duplicating logic.
