+++ src/server/routes/state-machine.ts
@@ -13,12 +13,12 @@ import type {
   FastifyRequest,
   FastifyReply,
 } from 'fastify';
-import { getDatabase } from '../db.js';
 import {
   STATES,
   STATE_MACHINE_TRANSITIONS,
 } from '../../shared/state-machine.js';
-import { DEFAULT_MESSAGE_TEMPLATES } from '../../shared/message-templates.js';
+import { getMessageTemplateService } from '../services/message-template-service.js';
+import { ServiceError } from '../services/service-error.js';
 
 // ---------------------------------------------------------------------------
 // Plugin
@@ -65,25 +65,13 @@ const stateMachineRoutes: FastifyPluginCallback = (
     '/api/message-templates',
     async (_request: FastifyRequest, reply: FastifyReply) => {
       try {
-        const db = getDatabase();
-        const dbTemplates = db.getMessageTemplates();
-        const dbMap = new Map(dbTemplates.map((t) => [t.id, t]));
-
-        // Merge defaults with DB overrides so every template is always returned
-        const enriched = DEFAULT_MESSAGE_TEMPLATES.map((def) => {
-          const dbRow = dbMap.get(def.id);
-          return {
-            id: def.id,
-            template: dbRow?.template ?? def.template,
-            enabled: dbRow?.enabled ?? true,
-            description: def.description,
-            placeholders: def.placeholders,
-            isDefault: !dbRow,
-          };
-        });
-
+        const service = getMessageTemplateService();
+        const enriched = service.listTemplates();
         return reply.code(200).send(enriched);
       } catch (err: unknown) {
+        if (err instanceof ServiceError) {
+          return reply.code(err.statusCode).send({ error: err.code, message: err.message });
+        }
         _request.log.error(err, 'Failed to get message templates');
         return reply.code(500).send({
           error: 'Internal Server Error',
@@ -103,40 +91,13 @@ const stateMachineRoutes: FastifyPluginCallback = (
         const { id } = request.params as { id: string };
         const body = request.body as { template?: string; enabled?: boolean } | null;
 
-        if (!body || (body.template === undefined && body.enabled === undefined)) {
-          return reply.code(400).send({
-            error: 'Bad Request',
-            message: 'Body must include at least one of: template, enabled',
-          });
-        }
-
-        const db = getDatabase();
-        const existing = db.getMessageTemplate(id);
-
-        if (existing) {
-          db.updateMessageTemplate(id, {
-            template: body.template,
-            enabled: body.enabled,
-          });
-        } else {
-          const defaultTmpl = DEFAULT_MESSAGE_TEMPLATES.find((t) => t.id === id);
-          if (!defaultTmpl) {
-            return reply.code(404).send({
-              error: 'Not Found',
-              message: `No message template found for id '${id}'`,
-            });
-          }
-
-          db.insertMessageTemplate({
-            id,
-            template: body.template ?? defaultTmpl.template,
-            enabled: body.enabled ?? true,
-          });
-        }
-
-        const updated = db.getMessageTemplate(id);
+        const service = getMessageTemplateService();
+        const updated = service.upsertTemplate(id, body ?? {});
         return reply.code(200).send(updated);
       } catch (err: unknown) {
+        if (err instanceof ServiceError) {
+          return reply.code(err.statusCode).send({ error: err.code, message: err.message });
+        }
         request.log.error(err, 'Failed to update message template');
         return reply.code(500).send({
           error: 'Internal Server Error',
